refactor: move components where they are used
